[move-execution] Move config out of runtime and verifier crates
Move `VMConfig`, `VMRuntimeLimitsConfig`, and `VerifierConfig` into
their own crate, separate from `move-vm/runtime` and
`move-bytecode-verifier` because these types are part of the interface
shared by all versions of `move-execution`, so cannot be duplicated
across multiple versions.

Test Plan:

```
external-crates$ sh ./tests.sh
sui$ cargo simtest
sui$ env SUI_SKIP_SIMTESTS=1 cargo nextest run
```
