+++ crates/workspace-hack/Cargo.toml
@@ -391,6 +391,7 @@ move-symbol-pool = { path = "../../external-crates/move/move-symbol-pool" }
 move-table-extension = { path = "../../external-crates/move/extensions/move-table-extension", default-features = false }
 move-transactional-test-runner = { path = "../../external-crates/move/testing-infra/transactional-test-runner", default-features = false }
 move-unit-test = { path = "../../external-crates/move/tools/move-unit-test", default-features = false }
+move-vm-config = { path = "../../external-crates/move/move-vm/config", default-features = false }
 move-vm-runtime = { path = "../../external-crates/move/move-vm/runtime", features = ["debugging", "testing"] }
 move-vm-test-utils = { path = "../../external-crates/move/move-vm/test-utils", features = ["tiered-gas"] }
 move-vm-types = { path = "../../external-crates/move/move-vm/types" }
@@ -1139,6 +1140,7 @@ move-symbol-pool = { path = "../../external-crates/move/move-symbol-pool" }
 move-table-extension = { path = "../../external-crates/move/extensions/move-table-extension", default-features = false }
 move-transactional-test-runner = { path = "../../external-crates/move/testing-infra/transactional-test-runner", default-features = false }
 move-unit-test = { path = "../../external-crates/move/tools/move-unit-test", default-features = false }
+move-vm-config = { path = "../../external-crates/move/move-vm/config", default-features = false }
 move-vm-runtime = { path = "../../external-crates/move/move-vm/runtime", features = ["debugging", "testing"] }
 move-vm-test-utils = { path = "../../external-crates/move/move-vm/test-utils", features = ["tiered-gas"] }
 move-vm-types = { path = "../../external-crates/move/move-vm/types" }
