+++ external-crates/move/move-bytecode-verifier/Cargo.toml
@@ -17,6 +17,7 @@ fail = "0.4.0"
 move-borrow-graph = { path = "../move-borrow-graph" }
 move-binary-format = { path = "../move-binary-format" }
 move-core-types = { path = "../move-core/types" }
+move-vm-config = { path = "../move-vm/config" }
 
 [dev-dependencies]
 hex-literal = "0.3.4"
