+++ external-crates/move/move-bytecode-verifier/bytecode-verifier-tests/src/unit_tests/code_unit_tests.rs
@@ -5,8 +5,9 @@
 use crate::support::dummy_procedure_module;
 use move_binary_format::file_format::Bytecode;
 use move_bytecode_verifier::meter::DummyMeter;
-use move_bytecode_verifier::{CodeUnitVerifier, VerifierConfig};
+use move_bytecode_verifier::CodeUnitVerifier;
 use move_core_types::vm_status::StatusCode;
+use move_vm_config::verifier::VerifierConfig;
 
 #[test]
 fn invalid_fallthrough_br_true() {
