+++ external-crates/move/move-bytecode-verifier/bytecode-verifier-tests/src/unit_tests/limit_tests.rs
@@ -5,11 +5,12 @@ use crate::unit_tests::production_config;
 use move_binary_format::file_format::*;
 use move_bytecode_verifier::{
     limits::LimitsVerifier, meter::DummyMeter, verifier::DEFAULT_MAX_IDENTIFIER_LENGTH,
-    verify_module_with_config_for_test, VerifierConfig,
+    verify_module_with_config_for_test,
 };
 use move_core_types::{
     account_address::AccountAddress, identifier::Identifier, vm_status::StatusCode,
 };
+use move_vm_config::verifier::VerifierConfig;
 
 #[test]
 fn test_function_handle_type_instantiation() {
