+++ external-crates/move/move-bytecode-verifier/src/stack_usage_verifier.rs
@@ -9,7 +9,7 @@
 //! the stack height by the number of values returned by the function as indicated in its
 //! signature. Additionally, the stack height must not dip below that at the beginning of the
 //! block for any basic block.
-use crate::{meter::Meter, VerifierConfig};
+use crate::meter::Meter;
 use move_binary_format::{
     binary_views::{BinaryIndexedView, FunctionView},
     control_flow_graph::{BlockId, ControlFlowGraph},
@@ -17,6 +17,7 @@ use move_binary_format::{
     file_format::{Bytecode, CodeUnit, FunctionDefinitionIndex, Signature, StructFieldInformation},
 };
 use move_core_types::vm_status::StatusCode;
+use move_vm_config::verifier::VerifierConfig;
 
 pub(crate) struct StackUsageVerifier<'a> {
     resolver: &'a BinaryIndexedView<'a>,
