+++ external-crates/move/move-vm/config/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "move-vm-config"
+version = "0.1.0"
+authors = ["Diem Association <opensource@diem.com>"]
+repository = "https://github.com/move-language/move"
+license = "Apache-2.0"
+publish = false
+edition = "2021"
+
+[dependencies]
+move-binary-format = { path = "../../move-binary-format" }
