+++ external-crates/move/move-vm/config/src/lib.rs
@@ -1,0 +1,5 @@
+// Copyright (c) The Move Contributors
+// SPDX-License-Identifier: Apache-2.0
+
+pub mod runtime;
+pub mod verifier;
