+++ external-crates/move/move-vm/runtime/Cargo.toml
@@ -22,6 +22,7 @@ smallvec = "1.6.1"
 
 move-bytecode-verifier = { path = "../../move-bytecode-verifier" }
 move-core-types = { path = "../../move-core/types" }
+move-vm-config = { path = "../config" }
 move-vm-types = { path = "../types" }
 move-binary-format = { path = "../../move-binary-format" }
