+++ sui-execution/latest/sui-adapter/Cargo.toml
@@ -19,6 +19,7 @@ move-binary-format.workspace = true
 move-bytecode-utils.workspace = true
 move-bytecode-verifier.workspace = true
 move-core-types.workspace = true
+move-vm-config.workspace = true
 move-vm-runtime.workspace = true
 move-vm-types.workspace = true
