+++ sui-execution/latest/sui-adapter/src/adapter.rs
@@ -6,13 +6,15 @@ use std::{collections::BTreeMap, sync::Arc};
 use anyhow::Result;
 use move_binary_format::{access::ModuleAccess, file_format::CompiledModule};
 use move_bytecode_verifier::meter::Meter;
-use move_bytecode_verifier::{verify_module_with_config_metered, VerifierConfig};
+use move_bytecode_verifier::verify_module_with_config_metered;
 use move_core_types::account_address::AccountAddress;
+use move_vm_config::{
+    runtime::{VMConfig, VMRuntimeLimitsConfig},
+    verifier::VerifierConfig,
+};
 pub use move_vm_runtime::move_vm::MoveVM;
 use move_vm_runtime::{
-    config::{VMConfig, VMRuntimeLimitsConfig},
-    native_extensions::NativeContextExtensions,
-    native_functions::NativeFunctionTable,
+    native_extensions::NativeContextExtensions, native_functions::NativeFunctionTable,
 };
 use sui_types::metrics::BytecodeVerifierMetrics;
 use sui_verifier::check_for_verifier_timeout;
