+++ sui-execution/latest/sui-verifier/Cargo.toml
@@ -12,6 +12,7 @@ move-binary-format.workspace = true
 move-bytecode-utils.workspace = true
 move-bytecode-verifier.workspace = true
 move-core-types.workspace = true
+move-vm-config.workspace = true
 
 sui-types = { path = "../../../crates/sui-types" }
 workspace-hack = { version = "0.1", path = "../../../crates/workspace-hack" }
