+++ sui-execution/latest/sui-verifier/src/meter.rs
@@ -2,11 +2,9 @@
 // SPDX-License-Identifier: Apache-2.0
 
 use move_binary_format::errors::{PartialVMError, PartialVMResult};
-use move_bytecode_verifier::{
-    meter::{Meter, Scope},
-    VerifierConfig,
-};
+use move_bytecode_verifier::meter::{Meter, Scope};
 use move_core_types::vm_status::StatusCode;
+use move_vm_config::verifier::VerifierConfig;
 
 struct SuiVerifierMeterBounds {
     name: String,
